Standalone liaison, add ability to disable built-in tabs
