+++ src/apps/zeromq/liaison/CMakeLists.txt
@@ -1,20 +1,40 @@
 add_definitions(-DGOBY_LIAISON_COMPILED_DOCROOT="${goby_SHARE_DIR}/goby/liaison")
 add_definitions(-DGOBY_LIAISON_INSTALLED_DOCROOT="${CMAKE_INSTALL_PREFIX}/share/goby/liaison")
 
-add_executable(goby_liaison
-  liaison.cpp
+add_library(goby_liaison_core SHARED
   liaison_home.cpp  
   liaison_wt_thread.cpp
   liaison_commander.cpp
   liaison_scope.cpp
+)
+
+add_executable(goby_liaison
+  liaison.cpp
+  )
+
+add_executable(goby_liaison_standalone
+  liaison.cpp
+  )
+
+target_compile_definitions(goby_liaison_standalone
+  PRIVATE LIAISON_STANDALONE=1
   )
 
 generate_middleware_interfaces(goby_liaison)
 
-target_link_libraries(goby_liaison
+set(LIB
   ${Boost_LIBRARIES}
   ${Wt_LIBRARIES}
   pthread
   goby
   goby_zeromq
+  goby_liaison_core
+)
+
+target_link_libraries(goby_liaison
+  ${LIB}
+  )
+
+target_link_libraries(goby_liaison_standalone
+  ${LIB}
   )
